Got logic error working
Previous error response now fixed with an exception. Can search in both strictly uppercase or lowercase. For loop is giving errors if you search the athlete that is not in its corresponding position in the list.
Eg. If you search bill or BILL, it will give you an error message 2x, then give you his stats, but if you searched bill, bill, paul, bill, you will result in 4x errors. This is because I am using a linear search and I've positioned the error message in a way that makes the code look really faulty.